+++ src/Tecan/Rack/DestLC.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+
+namespace MLL\Utils\Tecan\Rack;
+
+class DestLC extends BaseRack
+{
+    public function type(): string
+    {
+        return '96 Well MP LightCycler480';
+    }
+
+    public function name(): string
+    {
+        return 'DestLC';
+    }
+
+    public function positionCount(): int
+    {
+        return 96;
+    }
+}