+++ src/Tecan/Rack/DestTaqMan.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+
+namespace MLL\Utils\Tecan\Rack;
+
+class DestTaqMan extends BaseRack
+{
+    public function type(): string
+    {
+        return '96 Well PCR TaqMan';
+    }
+
+    public function name(): string
+    {
+        return 'DestTaqMan';
+    }
+
+    public function positionCount(): int
+    {
+        return 96;
+    }
+}