+++ tests/Tecan/CustomCommands/TransferWithAutoWashTest.php
@@ -6,7 +6,7 @@ use MLL\Utils\StringUtil;
 use MLL\Utils\Tecan\CustomCommands\TransferWithAutoWash;
 use MLL\Utils\Tecan\LiquidClass\CustomLiquidClass;
 use MLL\Utils\Tecan\Location\BarcodeLocation;
-use MLL\Utils\Tecan\Rack\CustomRack;
+use MLL\Utils\Tecan\Rack\FluidXRack;
 use PHPUnit\Framework\TestCase;
 
 final class TransferWithAutoWashTest extends TestCase
@@ -14,7 +14,7 @@ final class TransferWithAutoWashTest extends TestCase
     public function testTransferWithAutoWashCommand(): void
     {
         $liquidClass = new CustomLiquidClass('TestLiquidClassName');
-        $rack = new CustomRack('TestRackName', 'TestRackType');
+        $rack = new FluidXRack();
         $aspirateLocation = new BarcodeLocation('barcode', $rack);
         $dispenseLocation = new BarcodeLocation('barcode1', $rack);
 
@@ -22,8 +22,8 @@ final class TransferWithAutoWashTest extends TestCase
 
         self::assertSame(
             StringUtil::normalizeLineEndings(
-                'A;;;TestRackType;;barcode;100;TestLiquidClassName;;
-D;;;TestRackType;;barcode1;100;TestLiquidClassName;;
+                'A;;;96FluidX;;barcode;100;TestLiquidClassName;;
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;
 W;'
             ),
             $transfer->toString()